Moved 'setupUser' method to setup module
